# docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    description: model for user login
    properties:
      email:
        example: utilisateur@exemple.com
        type: string
      password:
        example: Motdepasse123
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  models.Sexe:
    enum:
    - MAN
    - WOMAN
    - OTHER
    type: string
    x-enum-varnames:
    - Male
    - Female
    - Other
...
  models.UserCreate:
    description: model for create a user
    properties:
      birthDayDate:
        example: "1990-01-01T00:00:00Z"
        type: string
      email:
        example: utilisateur@exemple.com
        type: string
      firstName:
        example: Jean
        type: string
      lastName:
        example: Dupont
        type: string
      password:
        example: Motdepasse123
        minLength: 6
        type: string
      sexe:
        allOf:
        - $ref: '#/definitions/models.Sexe'
        example: MAN
      userName:
        example: utilisateur123
        type: string
    required:
    - birthDayDate
    - email
    - firstName
    - lastName
    - password
    - sexe
    - userName
    type: object
...
host: localhost:8090
info:
  contact: {}
  description: API pour le projet PEC2 Backend
  title: API PEC2 Backend
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: user login with credential
      parameters:
      - description: User credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'token: JWT token'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'error: Wrong credentials or email not verified'
          schema:
            additionalProperties: true
            type: object
        "422":
          description: 'error: JWT not generated'
          schema:
            additionalProperties: true
            type: object
      summary: user login
      tags:
      - auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Create a new user with the provided information
      parameters:
      - description: User information
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.UserCreate'
      produces:
      - application/json
      responses:
        "201":
          description: 'message: User created successfully, email: user email'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: Invalid input'
          schema:
            additionalProperties: true
            type: object
        "409":
          description: 'error: Email already exists'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 'error: Error message'
          schema:
            additionalProperties: true
            type: object
      summary: Create a new user
      tags:
      - auth
...
  /valid-email/{token}:
    get:
      consumes:
      - application/json
      description: Resend validation email for users who loose their code or code
        is expired
      parameters:
      - description: user send email for received new mail
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 'message": "send email at user email address'
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 'error: User already validated account'
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 'error: User not found'
          schema:
            additionalProperties: true
            type: object
      summary: Resend Validation email
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: 'Entrez le JWT avec le préfixe Bearer: Bearer <JWT>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"